requirements.txt
gymnasium==1.0.0
numpy==2.2.4
pandas==2.2.3
matplotlib==3.10.1
python-dotenv==1.1.0
requests==2.32.3